refactor: format code and add config package
