+++ lib/App.bs.js
@@ -3,7 +3,7 @@
 
 var Log = require("./helpers/Log.bs.js");
 var ApiHome = require("./api/ApiHome.bs.js");
-var Routers = require("./helpers/Routers.bs.js");
+var Routers = require("./core/Routers.bs.js");
 var ApiUsers = require("./api/ApiUsers.bs.js");
 var PageHome = require("./pages/PageHome.bs.js");
 var PageAbout = require("./pages/PageAbout.bs.js");
