+++ lib/api/ApiHome.bs.js
@@ -2,7 +2,7 @@
 'use strict';
 
 var Json = require("@glennsl/bs-json/src/Json.bs.js");
-var Rest = require("../helpers/Rest.bs.js");
+var Rest = require("../core/Rest.bs.js");
 var Json_encode = require("@glennsl/bs-json/src/Json_encode.bs.js");
 
 function jsonResponse(c) {
