+++ lib/helpers/Log.bs.js
@@ -2,7 +2,7 @@
 'use strict';
 
 var List = require("bs-platform/lib/js/list.js");
-var Method = require("./Method.bs.js");
+var Method = require("../core/Method.bs.js");
 
 function transformTring(method_) {
   return String(method_);
