+++ lib/pages/PageHome.bs.js
@@ -1,7 +1,7 @@
 // Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
 'use strict';
 
-var Rest = require("../helpers/Rest.bs.js");
+var Rest = require("../core/Rest.bs.js");
 
 function run(param, response) {
   return Rest.bodyResponse(response, 200, "Bem vindo");
